+++ pages/_app.js
@@ -2,14 +2,15 @@ import "../styles/globals.css";
 import { ChakraProvider, InputLeftElement } from "@chakra-ui/react";
 import { QueryClientProvider } from "@tanstack/react-query";
 import { queryClient } from "../src/query";
-import { useEffect } from "react";
-import { getCookie } from "../src/cookie";
+import { AuthProvider } from "../src/context/AuthContext";
 
 function MyApp({ Component, pageProps }) {
   return (
     <ChakraProvider>
       <QueryClientProvider client={queryClient}>
-        <Component {...pageProps} />
+        <AuthProvider>
+          <Component {...pageProps} />
+        </AuthProvider>
       </QueryClientProvider>
     </ChakraProvider>
   );
