+++ pages/bets.js
@@ -21,9 +21,11 @@ import {
 import { useQuery } from "@tanstack/react-query";
 import { getBets } from "../src/api/api";
 import BetRow from "../src/components/BetRow";
+import { AuthContext } from "../src/context/AuthContext";
+import { useContext } from "react";
 
 export default function Bets() {
-  const user = queryClient.getQueryData(["user"]);
+  const { user } = useContext(AuthContext);
 
   const { data: bets } = useQuery({
     queryKey: ["bets"],
